requirements.txt
-e git+https://github.com/marcoplacenti/MLPipe.git@8465d61b17fcbc7eea6a110d80e6b10141e70fc3#egg=src
torch
torchvision
numpy
pandas
opencv-python
pillow
argparse
pyyaml
scikit-learn
wandb
uuid
pytorch-lightning
pytorch-lightning-bolts
ray[tune]
boto3
botocore
